wrap server with shared_ptr and improve precision of request time

+++ src/tinycore/asyncio/httpserver.h
@@ -194,7 +194,7 @@ public:
         return _protocol + "://" + _host + _uri;
     }
 
-    float requestTime() const;
+    double requestTime() const;
 
     const HTTPHeaders* getHTTPHeaders() const {
         return _headers.get();
